Entity refactoring
Entity refactoring
