+++ src/main/java/Cook/Cookify_SpringBoot/CookifySpringBootApplication.java
@@ -2,7 +2,12 @@ package Cook.Cookify_SpringBoot;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+import org.springframework.data.domain.AuditorAware;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+import java.util.Optional;
+import java.util.UUID;
 
 @EnableJpaAuditing
 @SpringBootApplication
@@ -13,6 +18,9 @@ public class CookifySpringBootApplication {
 		SpringApplication.run(CookifySpringBootApplication.class, args);
 	}
 
-
+	@Bean
+	public AuditorAware<String> auditorProvider() {
+		return () -> Optional.of(UUID.randomUUID().toString());
+	}
 }
