+++ src/main/java/Cook/Cookify_SpringBoot/domain/order/Order.java
@@ -1,6 +1,7 @@
 package Cook.Cookify_SpringBoot.domain.order;
 
-import Cook.Cookify_SpringBoot.domain.BaseTimeEntity;
+import Cook.Cookify_SpringBoot.global.Entity.BaseEntity;
+import Cook.Cookify_SpringBoot.global.Entity.BaseTimeEntity;
 import Cook.Cookify_SpringBoot.domain.delivery.Delivery;
 import Cook.Cookify_SpringBoot.domain.member.GoogleMember;
 import lombok.Getter;
@@ -12,7 +13,7 @@ import javax.persistence.*;
 @Table(name = "orders")
 @NoArgsConstructor
 @Entity
-public class Order extends BaseTimeEntity {
+public class Order extends BaseEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
